Fix: change Dni property type from double to long

+++ ShoppingCart.Application/Dtos/CartDto.cs
@@ -3,7 +3,7 @@
 public class CartDto
 {
     public int Id { get; set; }
-    public double UserDni { get; set; }
+    public long UserDni { get; set; }
     public List<CartItemDto> Items { get; set; } = new();
     public int TotalItems { get; set; }
     public decimal Total { get; set; }
